// lib/redux/types.ts
/**
 * Redux Action types
 */
export enum ActionType {
  SONG_LOAD_REQUESTED = "SONG_LOAD_REQUESTED",
  SONG_LOAD_COMPLETED = "SONG_LOAD_COMPLETED",

  USER_INFO_LOAD_REQUESTED = "USER_INFO_LOAD_REQUESTED",
  USER_INFO_LOAD_COMPLETED = "USER_INFO_LOAD_COMPLETED",

  USER_INFO_REMOVE = "USER_INFO_REMOVE",
  UPDATE_USER_LOCAL_DEVICE_TOKEN = "UPDATE_USER_LOCAL_DEVICE_TOKEN",
  UPDATE_TOKEN_IS_LOADING = "UPDATE_TOKEN_IS_LOADING",
  UPDATE_TOKEN_IS_VALID = "UPDATE_TOKEN_IS_VALID",

  ADD_FAVORITES_REQUESTED = "ADD_FAVORITES_REQUESTED",
  ADD_FAVORITES_COMPLETED = "ADD_FAVORITES_COMPLETED",

  REMOVE_FAVORITES_REQUESTED = "REMOVE_FAVORITES_REQUESTED",
  REMOVE_FAVORITES_COMPLETED = "REMOVE_FAVORITES_COMPLETED",

  LOAD_FAVORITES_REQUESTED = "LOAD_FAVORITES_REQUESTED",
  LOAD_FAVORITES_COMPLETED = "LOAD_FAVORITES_COMPLETED",

  UPDATE_EMBED_SONG_SHEET = "UPDATE_EMBED_SONG_SHEET",
  UPDATE_DISPLAY_SONG_TYPES = "UPDATE_DISPLAY_SONG_TYPES",
  UPDATE_SONG_SHEET_TYPE = "UPDATE_SONG_SHEET_TYPE",
  UPDATE_FAVORITES_SORT_BY = "UPDATE_FAVORITES_SORT_BY",
  UPDATE_DEVICE_SHOW_RELATED = "UPDATE_DEVICE_SHOW_RELATED",
  UPDATE_HYMN_CONTENT_SORT_BY = "UPDATE_HYMN_CONTENT_SORT_BY",

  AUDIO_PLAYER_QUEUE_SET = "AUDIO_PLAYER/queueSet",
  AUDIO_PLAYER_QUEUE_NEXT = "AUDIO_PLAYER/queueNext",
  AUDIO_PLAYER_SET_SONG = "AUDIO_PLAYER/setCurrSong",
  AUDIO_PLAYER_SET_DURATION = "AUDIO_PLAYER/setDuration",
  AUDIO_PLAYER_PLAY = "AUDIO_PLAYER/play",
  AUDIO_PLAYER_PAUSE = "AUDIO_PLAYER/pause",
  AUDIO_PLAYER_SET_TIME_PROGRESS = "AUDIO_PLAYER/setTimeProgress",
  AUDIO_PLAYER_REPEAT = "AUDIO_PLAYER/repeat",
  AUDIO_PLAYER_SHUFFLE = "AUDIO_PLAYER/shuffle",
  AUDIO_PLAYER_SHUFFLE_BOOK = "AUDIO_PLAYER/shuffleBook",
}